Pushing initial state to devtools
