+++ package.json
@@ -28,10 +28,12 @@
     "webpack": "^1.13.3"
   },
   "dependencies": {
-    "apollo-client": "^0.5.12",
+    "apollo-client": "^1.3.0",
     "graphql-syntax-highlighter-react": "^0.1.0",
     "graphql-tag": "^1.1.2",
     "lodash": "^4.17.2",
+    "prop-types": "^15.5.10",
+    "react-apollo": "^1.2.0",
     "react-redux": "^4.4.6",
     "uuid": "^3.0.1"
   }
